+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Fixed
 * Drop support for Ruby 2.6
 * Support Ruby 3.1, Rails 7.0
+* Replace Public Health England naming with NHS Digital
 
 ## 2.3.0 / 2022-01-14
 ### Fixed